// frontend/src/types/generated/workflows/conectaPROWorkflowsAPI.schemas.ts
export type WorkflowCategory = typeof WorkflowCategory[keyof typeof WorkflowCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WorkflowCategory = {
  CRM: 'CRM',
  HR: 'HR',
  FINANCE: 'FINANCE',
  OPERATIONS: 'OPERATIONS',
  ONBOARDING: 'ONBOARDING',
  MARKETING: 'MARKETING',
  SUPPORT: 'SUPPORT',
  COMMUNICATION: 'COMMUNICATION',
  DOCUMENT: 'DOCUMENT',
  INTEGRATION: 'INTEGRATION',
  MAINTENANCE: 'MAINTENANCE',
  SECURITY: 'SECURITY',
  ANALYTICS: 'ANALYTICS',
  CUSTOM: 'CUSTOM',
} as const;

export type WorkflowCreateDescription = string | null;

export type WorkflowCreateTags = string[] | null;

/**
 * Schema para criar workflow.
 */
export interface WorkflowCreate {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  description?: WorkflowCreateDescription;
  category?: WorkflowCategory;
  priority?: string;
  tags?: WorkflowCreateTags;
  tenant_id: string;
}

export type WorkflowResponseDescription = string | null;

export type WorkflowResponseTags = string[] | null;

/**
 * Schema de resposta para workflow.
 */
export interface WorkflowResponse {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  description?: WorkflowResponseDescription;
  category?: WorkflowCategory;
  priority?: string;
  tags?: WorkflowResponseTags;
  id: string;
  tenant_id: string;
  status: WorkflowStatus;
  version: number;
  total_executions: number;
  successful_executions: number;
  failed_executions: number;
}

/**
 * Status do workflow.
 */
export type WorkflowStatus = typeof WorkflowStatus[keyof typeof WorkflowStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WorkflowStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  PAUSED: 'PAUSED',
  ARCHIVED: 'ARCHIVED',
  ERROR: 'ERROR',
} as const;

export type WorkflowUpdateName = string | null;

export type WorkflowUpdateDescription = string | null;

export type WorkflowUpdateCategory = WorkflowCategory | null;

export type WorkflowUpdateStatus = WorkflowStatus | null;

export type WorkflowUpdatePriority = string | null;

export type WorkflowUpdateTags = string[] | null;

/**
 * Schema para atualizar workflow.
 */
export interface WorkflowUpdate {
  name?: WorkflowUpdateName;
  description?: WorkflowUpdateDescription;
  category?: WorkflowUpdateCategory;
  status?: WorkflowUpdateStatus;
  priority?: WorkflowUpdatePriority;
  tags?: WorkflowUpdateTags;
}

export type ListWorkflowsApiV1WorkflowsGetParams = {
/**
 * ID do tenant
 */
tenant_id: string;
status?: WorkflowStatus | null;
category?: WorkflowCategory | null;
/**
 * @minimum 0
 */
skip?: number;
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
};